#0: Fix static analyse, remove unused code; Implement more of cli docker client

+++ containers/containers-huskit/src/main/java/io/huskit/containers/cli/HtCli.java
@@ -3,22 +3,26 @@ package io.huskit.containers.cli;
 import io.huskit.common.Environment;
 import io.huskit.common.Nothing;
 import io.huskit.common.function.MemoizedSupplier;
+import io.huskit.containers.api.CliRecorder;
 import lombok.RequiredArgsConstructor;
 import lombok.SneakyThrows;
+import lombok.With;
 import lombok.experimental.NonFinal;
 import org.checkerframework.checker.nullness.qual.Nullable;
 
 import java.io.*;
 import java.util.ArrayList;
 import java.util.Objects;
+import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.function.Consumer;
 import java.util.function.Function;
-import java.util.function.Supplier;
 
 @RequiredArgsConstructor
 public class HtCli {
 
-    Supplier<DockerShellProcess> process = new MemoizedSupplier<>(this::createProcess);
+    @With
+    CliRecorder recorder;
+    MemoizedSupplier<DockerShellProcess> process = new MemoizedSupplier<>(this::createProcess);
 
     @SneakyThrows
     public <T> T sendCommand(HtCommand command, Function<CommandResult, T> resultConsumer) {
@@ -35,17 +39,26 @@ public class HtCli {
 
     @SneakyThrows
     private DockerShellProcess createProcess() {
-        return new DockerShellProcess();
+        return new DockerShellProcess(recorder);
+    }
+
+    @SneakyThrows
+    public void close() {
+        if (process.isInitialized()) {
+            process.get().stop();
+        }
     }
 
     static class DockerShellProcess {
 
         private static final String RUN_LINE_MARKER = "__HUSKIT_RUN_MARKER__";
         private static final String CLEAR_LINE_MARKER = "__HUSKIT_CLEAR_MARKER__";
+        private AtomicBoolean isStopped = new AtomicBoolean();
         private Process dockerProcess;
         private BufferedWriter commandWriter;
         private BufferedReader commandOutputReader;
         private BufferedReader commandErrorReader;
+        private CliRecorder recorder;
         @NonFinal
         @Nullable
         private String previousOutLine;
@@ -53,10 +66,11 @@ public class HtCli {
         @Nullable
         private String previousErrLine;
 
-        public DockerShellProcess() throws IOException {
-            ProcessBuilder builder = new ProcessBuilder().redirectErrorStream(true);
+        public DockerShellProcess(CliRecorder recorder) throws IOException {
+            this.recorder = recorder;
+            var builder = new ProcessBuilder().redirectErrorStream(true);
             if (Environment.is(Environment.WINDOWS)) {
-                builder.command("cmd");
+                builder.command("powershell");
             } else {
                 builder.command("/bin/sh");
             }
@@ -88,27 +102,34 @@ public class HtCli {
 
         @SneakyThrows
         private void doSendCommand(HtCommand command) {
+            recorder.record(command);
             clearBuffer();
             commandWriter.write(String.join(" ", command.value()));
             commandWriter.newLine();
             commandWriter.flush();
         }
 
+        @SneakyThrows
         public <T> T read(HtCommand command,
-                          Function<CommandResult, T> resultConsumer) throws IOException {
+                          Function<CommandResult, T> resultConsumer) {
+            var commandString = String.join(" ", command.value());
             var lines = new ArrayList<String>();
             var line = readOutLine();
             line = readOutLine();
             while (!line.endsWith(RUN_LINE_MARKER)) {
-                if (command.terminatePredicate().test(line)) {
-                    commandWriter.write("\003"); // ASCII code for "Ctrl+C"
-                    commandWriter.flush();
-                    clearBuffer();
-                    lines.add(line);
-                    break;
-                }
-                if (!line.endsWith(String.join(" ", command.value())) && command.linePredicate().test(line)) {
-                    lines.add(line);
+                if (!line.isEmpty()) {
+                    if (command.terminatePredicate().test(line)) {
+                        throw new UnsupportedOperationException();
+//                        commandWriter.write("\003"); // ASCII code for "Ctrl+C"
+//                        commandWriter.flush();
+//                        var ctrlc = Runtime.getRuntime().exec("powershell kill " + dockerProcess.pid());
+//                        ctrlc.waitFor();
+//                        lines.add(line);
+//                        break;
+                    }
+                    if (!line.endsWith(commandString) && command.linePredicate().test(line)) {
+                        lines.add(line);
+                    }
                 }
                 line = readOutLine();
             }
@@ -117,9 +138,11 @@ public class HtCli {
         }
 
         public void stop() throws IOException {
-            dockerProcess.destroyForcibly();
-            commandWriter.close();
-            commandOutputReader.close();
+            if (isStopped.compareAndSet(false, true)) {
+                dockerProcess.destroyForcibly();
+                commandWriter.close();
+                commandOutputReader.close();
+            }
         }
 
         private void clearBuffer() throws IOException {
@@ -127,7 +150,7 @@ public class HtCli {
             commandWriter.newLine();
             commandWriter.flush();
             var line = readOutLine();
-            while (Objects.equals(line, CLEAR_LINE_MARKER)) {
+            while (!Objects.equals(line, CLEAR_LINE_MARKER)) {
                 line = readOutLine();
             }
             previousOutLine = line;
@@ -137,9 +160,7 @@ public class HtCli {
         }
 
         private String readOutLine() throws IOException {
-            String line = commandOutputReader.readLine();
-            System.out.println(line);
-            return line;
+            return commandOutputReader.readLine();
         }
     }
 }
